requirements.txt
PyQt5>=5.15.0
python-nmap>=0.7.1